style(register): split in other files
