+++ backend/matcha/routes/__init__.py
@@ -8,6 +8,7 @@ import matcha.routes.interests
 import matcha.routes.notification
 import matcha.routes.pictures
 import matcha.routes.like
+import matcha.routes.register
 
 
 def init_routes(app):
@@ -19,3 +20,4 @@ def init_routes(app):
     app.register_blueprint(notification.bp)
     app.register_blueprint(pictures.bp)
     app.register_blueprint(like.bp)
+    app.register_blueprint(register.bp)
